+++ src/tools.mysql_drv/virtual_users.sql
@@ -1,7 +1,7 @@
-# $Id: virtual_users.sql,v 1.2 2009/05/23 18:28:57 sbajic Exp $
+# $Id: virtual_users.sql,v 1.21 2009/05/25 11:54:52 sbajic Exp $
 
 create table dspam_virtual_uids (
-  uid mediumint unsigned primary key AUTO_INCREMENT,
+  uid int unsigned primary key AUTO_INCREMENT,
   username varchar(128)
 ) type=MyISAM;
